+++ ms-reservation/src/main/java/com/eduardo/msreservation/dto/ReservationResponseDto.java
@@ -2,6 +2,7 @@ package com.eduardo.msreservation.dto;
 
 import com.eduardo.msreservation.enums.EStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -11,7 +12,7 @@ public record ReservationResponseDto(
         String userId,
         LocalDate checkInDate,
         LocalDate checkOutDate,
-        Double totalAmount,
+        BigDecimal totalAmount,
         EStatus status,
         String userEmail,
         String username
